fix(inference): diagnose blocked llama.cpp bridge
