+++ docs/inference/set-up-llama-cpp.mdx
@@ -165,6 +165,41 @@ Refer to [CLI Commands](../../reference/commands) for complete command behavior.
 
 ## Recover the Managed Runtime
 
+During managed installation, NemoClaw first checks internal runtime readiness and host-loopback health.
+It repeats these checks when onboarding resumes.
+It then runs the authoritative OpenShell Docker bridge probe on fixed port `8081`.
+If that bridge probe cannot connect or times out after the earlier checks pass, onboarding stops without changing UFW.
+The error reports these values:
+
+- The detected Docker network.
+- The source subnet.
+- The gateway IP address.
+- The fixed port `8081`.
+- The exact narrow UFW command.
+
+<Warning>
+The printed command changes UFW.
+Confirm that the reported Docker network, source subnet, and gateway IP address belong to the OpenShell Docker bridge before you run it.
+The rule allows TCP traffic only from that subnet to port `8081` on that gateway IP address.
+</Warning>
+
+Run the exact command from the onboarding error.
+It has this form:
+
+```bash
+sudo ufw allow from <subnet> to <gateway-ip> port 8081 proto tcp
+```
+
+After you apply the rule, rerun the same onboarding selection.
+
+If onboarding still reports the bridge failure, or when the managed runtime no longer needs the rule, remove the exact rule that you added:
+
+```bash
+sudo ufw --force delete allow from <subnet> to <gateway-ip> port 8081 proto tcp
+```
+
+Use the same subnet and gateway IP address that appeared in the original command.
+
 If onboarding stops, rerun the same provider and recipe selection.
 NemoClaw resumes only the exact persisted Docker authority, image, recipe, model digest, network, and runtime identity.
 It reconciles an unfinished create journal before it starts a new runtime.
